Move execution handle to the proper module
