+++ src/conductor/execution/ops/combine_outputs.py
@@ -4,10 +4,11 @@ from typing import Optional, Sequence, Tuple
 
 from conductor.context import Context
 from conductor.errors import CombineOutputFileConflict
+from conductor.execution.handle import OperationExecutionHandle
 from conductor.execution.ops.operation import Operation
 from conductor.execution.operation_state import OperationState
 from conductor.task_identifier import TaskIdentifier
-from conductor.task_types.base import TaskExecutionHandle, TaskType
+from conductor.task_types.base import TaskType
 
 
 class CombineOutputs(Operation):
@@ -34,7 +35,9 @@ class CombineOutputs(Operation):
     def main_task(self) -> Optional[TaskType]:
         return self._task
 
-    def start_execution(self, ctx: Context, slot: Optional[int]) -> TaskExecutionHandle:
+    def start_execution(
+        self, ctx: Context, slot: Optional[int]
+    ) -> OperationExecutionHandle:
         self._output_path.mkdir(parents=True, exist_ok=True)
 
         for dep_id, dep_dir in self._deps_output_paths:
@@ -57,8 +60,8 @@ class CombineOutputs(Operation):
             # The base data may be large, so we use symlinks to avoid copying.
             copy_into.symlink_to(relative_to_target)
 
-        return TaskExecutionHandle.from_sync_execution()
+        return OperationExecutionHandle.from_sync_execution()
 
-    def finish_execution(self, handle: TaskExecutionHandle, ctx: Context) -> None:
+    def finish_execution(self, handle: OperationExecutionHandle, ctx: Context) -> None:
         # Nothing special needs to be done here.
         pass
