+++ src/conductor/execution/ops/noop.py
@@ -1,10 +1,11 @@
 from typing import Optional
 
 from conductor.context import Context
+from conductor.execution.handle import OperationExecutionHandle
 from conductor.execution.ops.operation import Operation
 from conductor.execution.operation_state import OperationState
 from conductor.task_identifier import TaskIdentifier
-from conductor.task_types.base import TaskType, TaskExecutionHandle
+from conductor.task_types.base import TaskType
 
 
 class NoOp(Operation):
@@ -32,8 +33,10 @@ class NoOp(Operation):
     def main_task(self) -> Optional[TaskType]:
         return self._task
 
-    def start_execution(self, ctx: Context, slot: Optional[int]) -> TaskExecutionHandle:
-        return TaskExecutionHandle.from_sync_execution()
+    def start_execution(
+        self, ctx: Context, slot: Optional[int]
+    ) -> OperationExecutionHandle:
+        return OperationExecutionHandle.from_sync_execution()
 
-    def finish_execution(self, handle: TaskExecutionHandle, ctx: Context) -> None:
+    def finish_execution(self, handle: OperationExecutionHandle, ctx: Context) -> None:
         pass
